requirements.txt
flask==2.3.3
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.3.7
